+++ sprompt.cpp
@@ -70,7 +70,7 @@
 char *
 simple_prompt(const char *prompt, int maxlen, int echo)
 {
-	int			length;
+	size_t	   length;
 	char	   *destination;
 	FILE	   *termin,
 			   *termout;
@@ -149,7 +149,7 @@ simple_prompt(const char *prompt, int maxlen, int echo)
 	{
 		/* eat rest of the line */
 		char		buf[128];
-		int			buflen;
+		size_t		buflen;
 
 		do
 		{
